Update GTK+ to 2.20.1 and ATK to 1.29.2
This update is required to fix Korean IME related critical
bug(http://crosbug.com/18828). The reason of critical is that if this bug is
remained, we cannot input Hangul character into omnibox properly.

The main reason for this bug is conflicting preedit of Korean IME and Omnibox
inline suggestion from history, and it is hard to fix without GTK+ version up.
This bug does not represent in Linux, because Chrome on Linux works with
GTK+ >= 2.20.0. Such version of GTK+ support handling preedit(composing) state
from IME, and omnibox is avoid inline suggestion when user is composing.
Details are available at following codes.
 chrome/browser/ui/gtk/omnibox/omnibox_view_gtk.cc:614
 chrome/browser/ui/gtk/omnibox/omnibox_view_gtk.cc:837
I confirmed this update resolve this bugs(and also resolve another
bug.(http://crosbug.com/18827).

I think updating GTK+ is safe because in the case of Linux, chrome works fine
with GTK+ >= 2.20.1. Of course I confirmed this patch works fine on CR48 and
QEMU.

The ATK update is for dependency change of GKT+.

This CL is totally same as http://gerrit.chromium.org/gerrit/7586 .
I will announce this update to chromiums os dev group after submit this CL.

BUG=chromium-os:18828,chromium-os:18827
TEST=manually done

Change-Id: Ied3b54e931101fe6ef1cd0abbf0ab6a7f4a5a847
Reviewed-on: http://gerrit.chromium.org/gerrit/7760
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Seigo Nonaka <nona@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>