openssl: repair update_engine build.
Somewhere in the mists of time, <http://codereview.chromium.org/5552005> landed,
which added a stanza to the openssl config file which unconditionally requires
the pkcs11 engine, which is actually installed by opencryptoki. We don't install
opencryptoki on the host, so this config change actually broke openssl on the
host, but we did not notice because the openssl ebuild we had at that time was
masked for amd64, so on our hosts we did not use it.

Recently, I landed <http://gerrit.chromium.org/gerrit/7715>, which as a
side-effect unmasked our openssl ebuild on amd64. This caused openssl to get
built from source (as part of hard-host-depends) for the first time. On the
incremental builders (including trybots and the pfq), nothing seemed amiss
because portage's config protection refused updates to config files, but on the
full builders, the now-bogus config file prevented the host-side openssl from
running.

As a fix, introduce a 'pkcs11' USE flag to indicate that we should add pkcs11
support to the openssl config file, and use it for the target builds.

BUG=chromium-os:20060
TEST=Adhoc
emerge openssl on the host. grep 'pkcs11' /etc/ssl/openssl.cnf ; make sure there
are no entries.
emerge-$board openssl. grep 'pkcs11' /build/$board/etc/ssl/openssl.cnf ; make
sure there is a pkcs11_section.
emerge-$board update_engine should work again.

Change-Id: I033146b305ffa73af3a71ef8ec210ea7960d8613
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7769
Reviewed-by: David James <davidjames@chromium.org>