make_chroot: set up default list of locales
When upgrading glibc, if no locales are specified, it will generate all
of the available ones (which is over 400).  This can take quite a bit
of time and is (for the most part) wasted.  So have make_chroot set up
a small default list of locales that the buildbots need.

For existing chroots, we'll add an upgrade hook.  For new locales, the
enter_chroot script will generate things on the fly.

BUG=chromium-os:20378
TEST=delete chroot; run `cros_sdk`; review `locale -a` and /etc/locale.gen

Change-Id: I4e5a4c4b28d22d3be8fec8538cdd5ae42bc90372
Reviewed-on: http://gerrit.chromium.org/gerrit/7807
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>