Patch libpng-1.2.45 so that needed files are installed.
The upstream libpng-1.2.45 package only installs one .so file and
little else, apparently because it is intended to be installed only
as a companion to libpng-1.4.*.  This changelist patches the ebuild
file to do a full build/install so that the necessary files are
still installed.  A similar patch was done to 1.2.44, which has the
same issue.

BUG=chromium-os:18173
TEST=trybot on x86-generic-pre-flight-queue and arm-generic-bin

Change-Id: I7b61d60bc729537ffa6e2713503473f0e0160021
Reviewed-on: http://gerrit.chromium.org/gerrit/7880
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Matt Tennant <mtennant@chromium.org>