setup_board: key off tuples rather than arches for native builds
Rather than assuming all boards with a target arch of amd64 want to use
the native toolchain, do it only when the tuples are exactly the same.
Otherwise there would be collisions with the two toolchains.

This way we can install different dedicated tuples for amd64 boards and
get things wrapped the same as our other cross-compilers.

BUG=chromium-os:20636
TEST=run `./setup_board --board=amd64-generic` with toolchain.conf set to x86_64-cros-linux-gnu
TEST=`./setup_board --board=arm-generic && ./build_packages --board=arm-generic` still works
TEST=`cbuildbot chromiumos-sdk` still works

Change-Id: I6f6f4ffc40c87282261a7e82e392ec9dd478057a
Reviewed-on: http://gerrit.chromium.org/gerrit/8024
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Anush Elangovan <anush@chromium.org>