chromeos: undo sdk USE flag settings
At the moment, we do not have a dedicated profile for the cros sdk chroot.
Instead, we've been using things like profiles/arch/amd64/ to host all of
our settings.  This has worked fine except that it breaks amd64 boards.
So in the amd64 chromeos target profile, undo the arch/amd64/ settings.

This is meant as a short term solution.  Long term, we need to get the
cros sdk chroot onto its own dedicated profile so we can keep settings
specific it to it there without polluting the arch profiles.

BUG=chromium-os:20636
TEST=`emerge-amd64-generic openssh -pv` shows USE=-kerberos
TEST=`emerge openssh -pv` shows USE=kerberos

Change-Id: I34a7e69c6eae4c9daa8b85326f5cdaa4af9f6845
Reviewed-on: http://gerrit.chromium.org/gerrit/7997
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>