chromeos-base/chromeos-chrome: add amd64 support
The ebuild was lacking handling for the ARCH==amd64 case, so add the
little bit of glue necessary.  We also force the python arch as the
Chrome build logic assumes /usr/lib/ paths.  We'll get this fixed in
the upstream tree, but until then, force sanity for amd64 targets.

BUG=chromium-os:20636
TEST=`emerge-amd64-generic chromeos-chrome` now works
TEST=`emerge-x86-generic chromeos-chrome` still works
TEST=`emerge-arm-generic chromeos-chrome` still works

Change-Id: I631d5921c46096168623c905d4b085f3a82dba5b
Reviewed-on: http://gerrit.chromium.org/gerrit/8117
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>