x11-apps/xtrace: clean up ebuild
The src_compile step in EAPI=2 is not supposed to run configure, so
xtrace ends up running it twice.  So drop the src_compile func since
it is the same as the default.

When installing, we should be using `emake` so that we respect the
build MAKEOPTS settings as well as MAKE.  Installing in parallel is
awesome!

BUG=None
TEST=`emerge-amd64-generic xtrace` still works
TEST=`emerge-x86-generic xtrace` still works
TEST=`emerge-arm-generic xtrace` still works

Change-Id: I2f2a80911ce98814ac2133b1d1f251b2e27fd3f2
Reviewed-on: http://gerrit.chromium.org/gerrit/8198
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>