chromiumos-overlay: tcpdump: disable chroot
Turn off the chroot use flag when building tcpdump.  Because tcpdump
is installed on the stateful partition it cannot add the tcpdump user
and/or create the private directory for doing chroot'ed execution.
Since tcpdump is presently used only for network debugging and only
from a shell this is considered an acceptable tradeoff for now.  If
it ever moves to the base system we can turn it back on again.

BUG=chromium-os:20572
TEST=build a test image and verify tcpdump functions correctly

Change-Id: I53966cd1c7bce8f9d5d00bd4acb9b99a1bb17c08
Reviewed-on: http://gerrit.chromium.org/gerrit/8247
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>
Commit-Ready: Sam Leffler <sleffler@chromium.org>