chromeos-base/chromeos-chrome: rework arch<->feature logic
Unify all of the nacl checks into one helper func (use_nacl) so that we
don't have to copy & paste the logic in multiple places.

Drop the arch checks related to USE=chrome_pdf.  If we detect there is
a problem with an arch, we can package.use.mask the flag external to the
ebuild like we handle everything else.

For now, also add amd64 to the internal flash logic since we don't have
a dedicated USE flag for it.

BUG=chromium-os:20636
TEST=`emerge-amd64-generic chromeos-base/chromeos-chrome` now installs nacl
TEST=`emerge-x86-generic chromeos-base/chromeos-chrome` still works
TEST=`emerge-arm-generic chromeos-base/chromeos-chrome` still works

Change-Id: I682d6cb15f99b3db6724a1e50b3d61c4aceb00bb
Reviewed-on: http://gerrit.chromium.org/gerrit/8276
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>