Update Lohit fonts to the latest 2.5.0
The latest version of Lohit fonts are now available in OFL 1.1.

This update include the latest version of Lohit fonts for the following
scripts:

Devanagari, Bengali, Tamil, Telugu, Kannada, Malayalam, Gujarati,
Gurumukhi(Punjabi), Oriya

Lohit-Assamese is not included because it should be more or identical
to Bengali.

BUG=chromium-os:7447
TEST=Go to http://te.wikipedia.org and http://kn.wikipedia.org to see
if you don't see any more empty rectangles.

Change-Id: Iac64476096670f780dd20fd225fe8ffc678672df
Reviewed-on: http://gerrit.chromium.org/gerrit/8382
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Jungshik Shin (jungshik at google) <jshin@chromium.org>