setup_board: die early when host boards use --usepkg
The host boards will try to install normally and after ~45 minutes, fail
when installing the toolchains.  This is because they try to use binpkgs
which don't exist (since this board is the target which builds those in
the first place).  So reject attempts to run with --usepkg since it does
not really make any sense.

BUG=None
TEST=`./setup_board --board=amd64-host --usepkg` now fails early
TEST=`./setup_board --board=amd64-host --nousepkg` runs through normally

Change-Id: Ie24c4b5ca001e9aa55ccf8bb64f002d824cac94c
Reviewed-on: http://gerrit.chromium.org/gerrit/8646
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>