udev: unmask 171-r2 on amd64, arm & x86; update use flags
By umasking on the right architectures, and setting the use flags,
we can now just use the gentoo udev!

The primary change to the use flags is that "extras" has been split
into useful individual flags for the pieces we want, so we can just
specify "hwdb acl" to get those bits - and can turn off rule
generators which are on by default.

BUG=none
TEST=emerge, build_packages, build_image

Change-Id: Idb8abb5a38fd6e0e5870a69e6ab1b9d97983d6b0
Reviewed-on: http://gerrit.chromium.org/gerrit/8392
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>