udev: unmask 171-r2 on amd64, arm & x86; update use flags
By umasking on the right architectures, and setting the use flags,
we can now just use the gentoo udev!

The primary change to the use flags is that "extras" has been split
into useful individual flags for the pieces we want, so we can just
specify "hwdb acl gudev" to get those bits - and can turn off rule
generators which are on by default.

Also update autotest-deps and modemmanager (9999 and stable) to
expect the new USE flag.

BUG=none
TEST=emerge, build_packages, build_image

Change-Id: I59e964ca4afb9e25980e971888b8c1c4d4ce0ec2
Reviewed-on: http://gerrit.chromium.org/gerrit/8535
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>