amd64-target make.conf: do not clobber common USE
If we don't append USE, we clobber the flags that are set up in the
make.conf.common file.  This breaks the build when that file disables
critical USE flags.  This also brings the file in line with the other
targets which append USE.

BUG=chromium-os:20636
TEST=`emerge-amd64-generic atk` is no longer built with USE=introspection

Change-Id: Ic966b271e8bf8c3759a30f39a24747845eefe473
Reviewed-on: http://gerrit.chromium.org/gerrit/8485
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>