cros_set_lsb_release: use the FQDN for CHROMEOS_AUSERVER and CHROMEOS_DEVSERVER
$(hostname) returns just the host part of the devserver's domain name,
but this is not always enough to actually find the devserver when the
client device is not on the devserver's network or the network has a
crappy consumer cable modem that can't be configured correctly as the
DHCP server.

Use the FQDN of the devserver so that it can be resolved anywhere.

BUG=none
TEST=cat /etc/lsb-release, verify that the update/devserver's domain is fully-qualified

Change-Id: Ie12e7b9bf904d0c1e7a2d510f97ca1d20518a9b3
Reviewed-on: http://gerrit.chromium.org/gerrit/8626
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Aaron Plattner <aplattner@nvidia.com>
Tested-by: Darin Petkov <petkov@chromium.org>
Commit-Ready: Darin Petkov <petkov@chromium.org>