Make xf86-video-intel dependency on libXvMC conditional on USE flag.
Since we do not use XvMC, remove the dependency that was pulling libXvMC
and eselect-xvmc on x86 builds.  Do this by making the depdency in
xf86-video-intel conditional on a USE flag called 'xvmc'.

BUG=chromium-os:21093
TEST=trybot run of x86-generic-pre-flight-queue
Also build xf86-video-intel package on x86-generic and verify that the
libXvMC and eselect-xvmc packages are not built.

Change-Id: If295159186f377cc216f8e7a90afb377f5f72ca5
Reviewed-on: http://gerrit.chromium.org/gerrit/8656
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Matt Tennant <mtennant@chromium.org>