sys-devel/gcc: tweak multilib moving to handle our custom paths
The new gcc-4.6.0 build tweaks its version to be different from the $PV
value, so the assumption that the internal path is $PV fails.  Use the
same method that gcc does for getting at the internal version (run gcc
with the -dumpversion flag) to find out where things go.

Additionally, add a "|| die" so we can catch failures in the future
sooner, and make sure we don't try to move libs that are already in the
correct destination.

BUG=chromium-os:20636
TEST=`emerge cross-x86_64-cros-linux-gnu/gcc` and see libs in correct place
TEST=`emerge cross-i686-pc-linux-gnu/gcc` and see libs in same place
TEST=`emerge cross-arm-cros-linux-gnu/gcc` and see libs in same place
TEST=`./build_packages --board=amd64-generic` works again
TEST=`./build_packages --board=x86-generic` still works

Change-Id: Id785991a0a79c4e95319041ac8744ebc1706a158
Reviewed-on: http://gerrit.chromium.org/gerrit/8752
Reviewed-by: asharif <asharif@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>