Un-revert "Test a full merge without backtracking in build_packages."
If the latest version of a package can't be installed, build_packages
should fail so as to show the developer what they did wrong. This
ensures that changes that break the full build get caught earlier.

This ensures that anyone who runs build_packages will know when they
add a typo to an ebuild that prevents it from being merged. This is
important not just for developers, but also for trybots, the commit
queue, and the preflight queue.

This would have caught the breakage introduced by keybuk@ in
http://gerrit.chromium.org/gerrit/8378 where gpsd was used on arm
but the latest version was not marked as stable.

BUG=chromium-os:20981
TEST=Verify this catches bugs where latest version of package can't be
     emerged due to mismatched dependencies. Verify that all boards
     work with this option right now, including the public
     version of tegra2.

Change-Id: Ifcc270852aa4d511e370ee075babe7106fcc92b9
Reviewed-on: http://gerrit.chromium.org/gerrit/8827
Reviewed-by: Ryan Cui <rcui@chromium.org>
Tested-by: David James <davidjames@chromium.org>