wayland: import some ebuilds needed to build wayland
Not putting these in portage-stable because they are from the upstream
X11 overlay and not directly from the upstream portage.

Imported from:

git://git.overlays.gentoo.org/proj/x11.git

Not putting these into the wayland overland because wayland needs to
build a host tool in order to cross-compile itself.

BUG=none
TEST=none

Change-Id: Ic2183e8a1026febb25c3e81948f21461cb90f746
Reviewed-on: http://gerrit.chromium.org/gerrit/8937
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Mandeep Singh Baines <msb@chromium.org>