update flashrom ebuild and package.use files
This patch modernizes the Flashrom ebuild to take advantage of improvements
made upstream.

The major improvement here is that we can employ USE flags to control CONFIG_*
Makefile variables. This allows us to move the programmer support logic to
architecture/board-specific package.use and package.use.mask files.

Programmer support is as follows:
- The host should have all USB and serial programmers made available for
  things like Servo, Dediprog, etc. Any extra programmers added for this are
  disabled for targets.

- ARM targets must not use any PCI-centric programmer due to libpci issues
  (see http://patchwork.coreboot.org/patch/3255/ for details).

BUG=chromium-os:21405
TEST=Tested by building flashrom and reading firmware on all supported ARM and x86 targets.

Change-Id: I9e0e41101d70cc2b630b116685500f621e5a4d1e
Reviewed-on: http://gerrit.chromium.org/gerrit/9760
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>