sys-kernel/chromeos-kernel{,-next}: clean up ebuilds a bit
These ebuilds duplicate a lot of the same constructs in multiple places
(like updating the config and running make), so add local helpers to keep
all the duplicate logic in one place.

This also allows us to localize a lot of the code in global scope to the
right function where it actually gets used (and in some cases, delete the
variables altogether).

Finally, the most important piece here, we sanitize the LDFLAGS value that
is coming from the build env to a value the kernel build can utilize.

Note: this contains a slight change to fix the regressions introduced by
the previous attempt where a refactor of the toolchain settings enabled
the gold workarounds for all boards.

BUG=none
TEST=`emerge-amd64-generic chromeos-kernel` still works
TEST=`emerge-arm-generic chromeos-kernel` still works
TEST=`emerge-x86-alex chromeos-kernel` still works
TEST=`emerge-x86-generic chromeos-kernel` still works
TEST=`emerge-x86-pineview chromeos-kernel` still works
TEST=`emerge-tegra2_seaboard chromeos-kernel` still works
TEST=`emerge-amd64-generic chromeos-kernel-next` still works
TEST=`emerge-arm-generic chromeos-kernel-next` still works
TEST=`emerge-x86-alex chromeos-kernel-next` still works
TEST=`emerge-x86-generic chromeos-kernel-next` still works
TEST=`emerge-x86-pineview chromeos-kernel-next` still works

Change-Id: Ic12edf42922c42fbbb151bbd25abd3648f6c74ea
Reviewed-on: http://gerrit.chromium.org/gerrit/9881
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Chris Wolfe <cwolfe@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>