sys-kernel/chromeos-kernel{,-next}: clean up ebuilds a bit
These ebuilds duplicate a lot of the same constructs in multiple places
(like updating the config and running make), so add local helpers to keep
all the duplicate logic in one place.

This also allows us to localize a lot of the code in global scope to the
right function where it actually gets used (and in some cases, delete the
variables altogether).

Finally, the most important piece here, we sanitize the LDFLAGS value that
is coming from the build env to a value the kernel build can utilize.

BUG=none
TEST=`emerge-amd64-generic chromeos-kernel` still works
TEST=`emerge-x86-generic chromeos-kernel` still works
TEST=`emerge-arm-generic chromeos-kernel` still works
TEST=`emerge-amd64-generic chromeos-kernel-next` still works
TEST=`emerge-x86-generic chromeos-kernel-next` still works
TEST=`emerge-arm-generic chromeos-kernel-next` still works

Change-Id: I284de46433b1198478e5441e7b389482faf33827
Reviewed-on: http://gerrit.chromium.org/gerrit/9772
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>