laptop-mode-tools: hdparm check for valid drive
Sometimes the regexes get interpreted as literals, when no matches are
found.  So it can result in those regexes being used as drive names,
which will result in an error.  The error is nonfatal but it prints out
some logging.  This patch cleans up that logging by recognizing the case
of an invalid drive name, and indicating that it will not be checked.

BUG=chromium-os:21045
TEST=Run:
1. turn on laptop_mode debugging (set VERBOSE_OUTPUT=1 in
/etc/laptop-mode/laptop-mode.conf)
2. sudo /usr/sbin/laptop_mode auto force
3. Check messages.  You should not see something like:
     Querying /dev/mmcblk[0-9]p[0-9] media type using udevinfo:
     device node not found

Change-Id: I75716326a328f5c54fde3113f3a6b16ce3334735
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/9880
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>