Move u-boot-scripts from overlay-tegra2.
These script are useful on any system using u-boot (the "legacy"
version of u-boot, specifically).  There is nothing tegra2 about them.

There are no changes between this and v4 (which lived in
overlay-tegra2).  The rev in version number just makes it easy to see
that a build has picked up the files from the new location.

A followup change will be made to delete the copy in overlay-tegra2.

BUG=chromium-os:21540
TEST=Made sure that 'emerge-${BOARD} u-boot-scripts' built this new
version (r5).
TEST=Made sure that u-boot-scripts is now a dependency of chromeos.

Change-Id: I1bb40db3308f2412e364010e647e0cb91e6c5e80
Reviewed-on: http://gerrit.chromium.org/gerrit/9874
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Dave Parker <dparker@chromium.org>