sys-libs/glibc: pull cross updates from upstream
The current logic always moves around the glibc lib dirs instead of
respecting the multilib env.  Pull down fixes from upstream for this.

We also need to update the package.mask settings as it prevents us
from revbumping the current 2.11.1 ebuilds.  Move it to 2.12 so we
can expand 2.11 as much as we want.  And slip in a missing mask for
the new x86_64-cros-linux-gnu toolchain.

BUG=chromium-os:21387
TEST=`crossdev -C x86_64-cros-linux-gnu && crossdev x86_64-cros-linux-gnu` uses /lib64/ subdir
TEST=`crossdev -C i686-pc-linux-gnu && crossdev i686-pc-linux-gnu` uses /lib/ subdir
TEST=`cbuildbot chromiumos-sdk` works

Change-Id: I1c13e2f32c16e67f04e939cb0129d19471b04395
Reviewed-on: http://gerrit.chromium.org/gerrit/9941
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>