fontconfig: Reduce optimization level
This fixes a failure to generate the fontconfig cache for ARM targets.
Unfortunately the cache is platform-specific, so fc-cache is run
under qemu-arm. A recent change resulted in the -O2-compiled fc-cache
causing a segfault and failing in some scenarios.

Because this reduces the optimization level on the entire fontconfig
package it is a temporary fix.

BUG=chromium-os:21493
TEST=Manual emerge-arm-generic chromeos no longer fails on qemu-arm.
  cbuildbot for x86 and arm generates a valid cache.

Change-Id: I1bc8e3dbbc6896c0ea9881a6557fe4d085b857da
Reviewed-on: http://gerrit.chromium.org/gerrit/10014
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>