chromeos-test: add dnsmasq to RDEPEND (x86-only)
we need a dhcp server in test images, so we can run some of the WiFiManager
autotests in a VM. previously, we used the server in the net-misc/dhcp
package. but that package is now being built without the server component
(see http://gerrit.chromium.org/gerrit/8478).

resolve the problem by adding dnsmasq to the test image dependencies, and
using it for the DHCP server in the tests.

BUG=chromium-os:21724
TEST=manual: emerge chromeos-test, verify dnsmasq is present

Change-Id: Idf46f10da2bd3145d4df3d2aae2ea9a0b57671a7
Reviewed-on: http://gerrit.chromium.org/gerrit/10179
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>