sci-geosciences/gpsd: hard depend on python
The gpsd package uses python to build/install files, and its own
configure script will abort if python is not found.  So make the
dependency unconditional.  This matches upstream Gentoo.

Making the python dep conditional doesn't seem to be much of a
gain (if any), and it doesn't look trivial.

BUG=None
TEST=`./build_packages && ./build_image` for x86-alex adds no new deps

Change-Id: Ifc1ba8009bc0997e9807f7d05175892d13fe4c44
Reviewed-on: https://gerrit.chromium.org/gerrit/10306
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>