Add USE=aura support to chromeos-chrome and chromeos-login
Aura requires that a GYP_DEFINE (use_aura) be set for the
chromeos-chrome build and that the window manager is disabled if Aura
is built.  This change adds a USE flag to the packages necessary for
this to happen.

This CL must go in immediately before
http://gerrit.chromium.org/gerrit/10425.

CRBUG=http://crbug.com/100628
BUG=None
TEST=Ran cbuildbot against x86-generic
  Ran build_packages with USE=aura, confirmed GYP flags
  Deployed image to device with "standard" flags (USE=-aura) and confirmed

Change-Id: I5ae297ed79f64b977fa143c9cac6b3bf4373d2b4
Reviewed-on: http://gerrit.chromium.org/gerrit/10424
Tested-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: Jon Kliegman <kliegs@chromium.org>