depend on both new and old shflags packages
We are moving to a new shflags package which comes from upstream Gentoo.
It installs into a different location, so we can have both installed at
the same time.  The overhead is trivial, and this is just transitional.

Once the relevant packages have their source code updated to use the new
shflags install location, we can drop the old dep.  We do this first so
that things don't have a race condition chance to fail.

BUG=chromium-os:21742
TEST=build_packages+build_image --test for x86-alex boots up fine
TEST=gmerge still works

Change-Id: I8144b717737ac3ef5f7a3ec26214a24bd44698e2
Reviewed-on: http://gerrit.chromium.org/gerrit/10300
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>