chromeos-base/chromeos-base: handle /etc/localtime when USE=-cros_host
The /etc/localtime symlink is unconditionally installed.  While we want
this behavior for the boards, we don't want this for the cros sdk as we
manually setup that file from the user's external system.  So add the
cros_host USE flag to control the behavior.

BUG=chromium-os:21850
TEST=`emerge chromeos-base` no longer clobbers /etc/localtime

Change-Id: I4f16421acad6cdc70e39ad226dd90e8cb2417c69
Reviewed-on: https://gerrit.chromium.org/gerrit/10960
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>