app-shells/[bd]ash: move /bin/sh management to chromeos-base/chromeos-base
The only reason we have bash/dash in the cros overlay is so that we have
the /bin/sh symlink pointing to bash in the cros sdk, but pointing to dash
in the board root.  Once we move that handling to chromeos-base, we can
drop all the bash/dash ebuild customizations.

BUG=chromium-os:21901
TEST=`emerge chromeos-base bash` and see /bin/sh -> bash
TEST=`emerge-amd64-generic chromeos-base dash` and see /bin/sh -> dash
TEST=`emerge-arm-generic chromeos-base dash` and see /bin/sh -> dash
TEST=`emerge-x86-generic chromeos-base dash` and see /bin/sh -> dash
TEST=`cbuildbot arm-generic-full` passes

Change-Id: I1a6e632d561b02c1ad906be0432552c45efdf566
Reviewed-on: https://gerrit.chromium.org/gerrit/10963
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>