Teach setup_board to grab boto files from overlays.
This is needed to allow files to be fetched from Google Storage
using custom boto files. We already have some configuration for this
in the various make.conf files, but we don't currently have the ability
to override board-specific configuration in chromeos-overlay.

This CL adds that capability, by simply defining the appropriate fetch
commands for all boards.

BUG=chromium-os:21977
TEST=Verify that FETCHCOMMAND_GS is setup correctly for all overlays, so
     everybody can download their prebuilts still. Also verify that the
     new boto file in chromeos-overlay allows folks to access a superset
     of all of the prebuilts. Trybot run.

Change-Id: I470e4f2fb6d1ed9b59f4c49a5d7979b42fdcc1ef
Reviewed-on: http://gerrit.chromium.org/gerrit/10548
Reviewed-by: Anush Elangovan <anush@google.com>
Commit-Ready: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>