Add IxChariot Performance Endpoint to test images
In the chromeos-test package, add a dependency on ixchariot.  Added a
placeholder ixchariot ebuild that does nothing.  Real ixchariot ebuild files
in a private overlay.

BUG=chromium-os:21991
TEST=Built Tegra2 image, which does not have a ixchariot package in its priv
overlay, and verified that it builds OK.  Built a x86-zgb, which does have a
ixchariot ebuild in a private overlay, and veriied that the package from the
private overlay is included in the build.

Change-Id: I01387b9063c42de48416365cb8b5d3772c036f09
Reviewed-on: https://gerrit.chromium.org/gerrit/10596
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>