Mark as stable: "Update bots in golo.chromium.org to use appropriate mirror."
webkit.org has asked us to stop downloading webkit from our bots, and instead
to use a mirror. This mirror is only accessible from the golo -- hence
we're only updating the golo bots to use it.

This change propagates the change to the stable ebuild as well so that we
don't have to wait for Chrome uprev.

BUG=chromium-os:17959
TEST=Run this from inside the golo.

Change-Id: Ic068602c1910baba5627970732d36543837f48b7
Reviewed-on: https://gerrit.chromium.org/gerrit/10757
Reviewed-by: Anush Elangovan <anush@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/10768
Reviewed-by: David James <davidjames@chromium.org>