media-libs/pico: overhaul [e]build
This cleans up the ebuild to use general Gentoo style in ebuild writing.
This implicitly fixes issues related to the right ABI libdir for libpico.a

It also simplifies the local Makefile used to build things so that we
can build in parallel, and it runs ranlib on the archive to generate a
nice index.

BUG=chromium-os:20636
TEST=`emerge-amd64-generic pico` now installs into /usr/lib64
TEST=`emerge-arm-generic pico` installs into same places
TEST=`emerge-x86-generic pico` installs into same places

Change-Id: I6055a877678526410e1ce81e613068cd16e8b79e
Reviewed-on: https://gerrit.chromium.org/gerrit/11006
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>