python: export cached paths for autotools
Our current make.conf code hardcodes the /lib/ subdir and the 2.6 version
for some python paths.  We also have custom python patches to guess at the
right path to return when asked.

Unfortunately, the latest automake code for querying python does so in a
way that is indistinguishable from running outside of the autotool env.
And the information given to python is not enough for it to accurately
deduce the right libdir path.

For now, add a bashrc hook which sets up both of these python path vars
based on the active python.  Since we are exporting variables that only
autoconf will use, we can assume that '${libdir}' exists, and so we use
that to automatically get the right libdir path.  When ebuilds run econf,
they pass --libdir to the right place.

We also need to update pygtk and xcb-proto ebuilds to include libdir= in
their .pc files so we can base pyexecdir off of it.

BUG=chromium-os:22413
TEST=`emerge python; emerge-amd64-generic dbus-python` and see dbus-python in /usr/lib64/
TEST=`emerge python; emerge-x86-generic dbus-python` and see dbus-python in /usr/lib/
TEST=`cbuildbot amd64-generic-full` works
TEST=`cbuildbot x86-generic-full` works

Change-Id: I071b3ce1e0880c5ee9ff8ae2a54a40c47cad20c8
Reviewed-on: https://gerrit.chromium.org/gerrit/11066
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>