Add serial-tty build
This is based on the existing tegra-debug build (which should now
be deleted).

The old tegra-debug build was only installing the ttyS0.conf file
(plus a device file that was no longer needed), so the name
tegra-debug didn't make sense.  ...and by putting it here, other
boards can use it (and devs can gmerge it on x86 if they have a serial
port).

BUG=chromium-os:22507
TEST=Built and saw ttyS0.conf show up.

Change-Id: I26050be5b7d4da0106c834f0772a0c69ab4ccb81
Reviewed-on: https://gerrit.chromium.org/gerrit/11164
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>