curl-7.21.4-r2: Ensure that interfaces names are not looked up via DNS
Ensure that interfaces names are not looked up via DNS because doing
so blocks a thread inside curl, even if it is using c-ares for name
resolution.  If a user has passed an interface name the intention is
not lookup the name via DNS if perchance the interface has
disappeared.

Test this by running strace on flimflam while running
network_3GModemControl. ensure no polls statements with a single fd
and timeout of 1000.

BUG=chromium-os:22240
TEST=see above.

Change-Id: I7943b6665ddaf9c9aa98a855e04b4b8ba49c2b9a
Reviewed-on: https://gerrit.chromium.org/gerrit/11186
Commit-Ready: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>