wrap old style config scripts
Not all packages have converted to pkg-config/.pc, so make sure all the
old xxx-config scripts have proper $CTARGET-xxx-config wrappers.  If we
don't, we end up using the config script from the host chroot, which will
often pull in -I/-L paths pointing to the chroot and not the board root.

We can't just symlink the wrappers themselves as we have a one-to-many
mapping for ctarget-to-sysroot, so we add a small generic wrapper that
first figures out what script it is wrapping, then combines that with the
active SYSROOT to ultimately locate the target xxx-config.

Since the board xxx-config doesn't contain paths to the board sysroot,
we mung the output to make it point to the right place on the fly.

To get the symlinks to wrap the xxx-config script in the first place,
we add some pkg_pre/postinst hooks to create things on the fly.

This isn't a short term hack to make things work until we have a better
solution in place.  The only way to "fix" this long term is to get the
respective upstream to migrate their package to pkg-config rather than
using an old style xxx-config script.

BUG=chromium-os:21276
TEST=build_packages for arm-generic now passes w/out -I/-L hacks
TEST=`cbuildbot arm-generic-full` passes
TEST=`cbuildbot amd64-generic-full` passes
TEST=`cbuildbot x86-generic-full` passes

Change-Id: I7342d9347bf8e22071cccbeca708e95f975fe86b
Reviewed-on: https://gerrit.chromium.org/gerrit/11626
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>