Managing caret position by mozc-engine.
We want to fix crosbug.com/21356 but we do not have API to get composition
boundary rectangle. So this patch is workaround for them. The mozc-server
watching current caret position and if state change from precomposition to
composition is happend, the mozc-server regards as compsition begining position.

BUG=chromium-os:21356
TEST=mozc-side test passed and manually test on Cr48.

Change-Id: I066ae40996e956ccda3eca499839e51528db951c
Reviewed-on: https://gerrit.chromium.org/gerrit/11465
Reviewed-by: Hiroyuki Komatsu <komatsu@chromium.org>
Tested-by: Seigo Nonaka <nona@chromium.org>
Commit-Ready: Seigo Nonaka <nona@chromium.org>