Update perf and perf-next ebuilds to be closer to upstream
This fixes the problems with perf being compiled with a perl
dependency which causes it not to work on amd64-generic.
Updating to EAPI4 means we don't need to put "|| die" after every
command.  I also renamed the "tui" USE flag to "ncurses" which
matches what is used upstream.  Finally, it also installs the
perf-archive command.

BUG=chromium-os:21388
BUG=chromium-os:13831
TEST=emerge-amd64-generic perf; make sure it runs on amd64-generic
TEST=emerge-amd64-generic perf-next; make sure it runs on amd64-generic
TEST=emerge-x86-generic perf; make sure it runs on x86-generic
TEST=emerge-x86-generic perf-next; make sure it runs on x86-generic
TEST=emerge-arm-generic perf; make sure it runs on arm-generic
TEST=emerge-arm-generic perf-next; make sure it runs on arm-generic
TEST=emerge perf; make sure it runs inside the chroot (perl and python)
TEST=emerge perf-next; make sure it runs inside the chroot (perl and python)

Change-Id: I7f481bd3d1556a13b49ef1c873b8d876418e1b03
Reviewed-on: https://gerrit.chromium.org/gerrit/11518
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>